Week 2/workflows/prefect/requirements.txt
jupyter
prefect
pandas
numpy
mlflow
xgboost
sqlalchemy
pymysql
scikit-learn
pyyaml
dynaconf
